feat(xeps): check server-supported XEPs and put them in a store
XEPs supported by the server are located in a new 'xeps' store as an object mapping XEP-CODE to
boolean. Natural language data about each XEP is also provided. Checks that work on disco results
are secluded in their own file, while checks that happen when receiving <stream:features/> are in
the stanza.js plugin.

BREAKING CHANGE: KNOWN ISSUE: XEP-0352 is missing in the 'features' event in stanza.js, meaning this
XEP will never be detected